feat: update remove result service to try to find result before to delete it

+++ src/services/remove-result.ts
@@ -7,6 +7,10 @@ export class RemoveResult {
   }
 
   async execute (id: string): Promise<void> {
+    const resultInDb = await this._repository.findOne(Number(id))
+    if (resultInDb === null) {
+      throw new Error()
+    }
     await this._repository.delete(id)
   }
 }
